Исправить замечания в ФЗ8
